requirements.txt
Flask==1.0.2
Flask-Cors==3.0.7
uWSGI==2.0.17.1
PyMySQL==0.9.2
SQLAlchemy==1.2.14
Flask-Migrate==2.3.0
python-dotenv==0.9.1
Flask-Bcrypt==0.7.1
Flask-Script==2.0.6
PyJWT==1.6.4
